\name{DispatchMethods}
\alias{DispatchMethods}
\non_function{}
\title{S-level constants}
\description{
Constants used to communicate between S and C code
to indicate how a COM call
should be invoked.
This can be used as the value of the \code{.dispatch}
argument in the \code{\link{.COM}}
function to control whether
to treat the call 
as a property accessor, method call
or both.
}
\usage{data(DispatchMethods)}
\format{

}
\source{
}
\references{
\url{http://www.omegahat.net/RDCOMClient}
\url{http://www.omegahat.net/RDCOMServer}
\url{http://www.omegahat.net/SWinTypeLibs}
\url{http://www.omegahat.net/SWinRegistry}
}

\examples{
\dontrun{

}
}
\keyword{datasets}
